GridModelStream works with integer types

+++ harness/resources/res/test.vert
@@ -1,6 +1,6 @@
 @@Version
 
-@@In pos = vec3
+@@In pos = ivec3
 @@In normal = vec3
 
 @@Passthrough uv = vec2
